# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnHub learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{clerkId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /users/me:
    get:
      operationId: getMe
      tags: [admin]
      summary: Get current user profile and subscription info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [admin]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, userId, status, plan, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        esewaTransactionId: { type: ["string", "null"] }
        status: { type: string }
        plan: { type: string }
        currentPeriodEnd: { type: ["string", "null"], format: date-time }
        cancelAtPeriodEnd: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [clerkId, email, name, role, createdAt]
      properties:
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        subscriptionStatus: { type: ["string", "null"] }
        subscriptionPlan: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [user, admin] }
    UserProfile:
      type: object
      required: [clerkId, email, name, role, createdAt]
      properties:
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        subscription:
          $ref: "#/components/schemas/Subscription"
        createdAt: { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        name: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }

    # ─── Videos ───────────────────────────────────────────────────────────────
...